\name{AMZN}
\docType{data}
\alias{AMZN}
\title{Log return of Amazon inc. closing Value}
\description{
  The Amazon inc. closing value log return from 1998-01-01 to 2015-12-31 from
  Yahoo Finance \url{https://finance.yahoo.com/}.
}
\usage{data("AMZN")}
\format{Matrix containing 4529 observations.}
\source{Yahoo Finance \url{https://finance.yahoo.com/}}
\keyword{datasets}
