% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.R
\name{BIC}
\alias{BIC}
\title{Compute Bayesian information criterion (BIC).}
\usage{
BIC(fit)
}
\arguments{
\item{fit}{Fit object of type \code{\link{MSGARCH_MLE_FIT}} created with \code{\link{fit.mle}} or \code{\link{MSGARCH_BAY_FIT}} created with \code{\link{fit.bayes}}}
}
\value{
BIC value.
}
\description{
Compute Bayesian information criterion (BIC).
}
\details{
If a matrix of MCMC posterior draws estimates is given, the BIC on the posterior mean is calculated.
}
\examples{
data("sp500ret")

spec = MSGARCH::create.spec(model = c("sGARCH","sGARCH"), distribution = c("norm","norm"),
                             do.skew = c(FALSE,FALSE), do.mix = FALSE, do.shape.ind = FALSE) 
                             
fit = MSGARCH::fit.mle(spec = spec, y = sp500ret)

BIC = MSGARCH::BIC(fit = fit)
}
\references{
Schwarz, G. (1978). Estimating the dimension of a model. \emph{Annals of Statistics}, 6, pp. 461-464.
}

