% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSGARCH.R
\docType{data}
\name{dem2gbp}
\alias{dem2gbp}
\title{DEM/GBP exchange rate log-returns}
\format{
\code{vector} of size 1,974.
}
\usage{
data("dem2gbp")
}
\description{
The \code{vector} \code{dem2gbp} contains daily observations of the Deutschmark vs British Pound foreign exchange
rate log-returns. This dataset has been promoted as an informal benchmark for GARCH
time-series software validation. See McCullough and Renfro (1999), and Brooks, Burke, and Persand
(2001) for details. The nominal returns are expressed in percent as in Bollerslev and Ghysels
(1996). The sample period is from January 3, 1984, to December 31, 1991, for a total of 1974
observations.
}
\references{
Bollerslev T., Ghysels, E. (1996)
Periodic autoregressive conditional heteroscedasticity.
\emph{Journal of Business and Economic Statistics}, 14, 139-151.

Brooks C., Burke S. P., Persand G. (2001)
\emph{International Journal of Forecasting}, 17, 45-57.
\url{http://doi.org/10.1016/S0169-2070(00)00070-4}

McCullough B. D., Renfro C. G. (1999)
Benchmarks and software standards: A case study of GARCH procedures. 
\emph{Journal of Economic and Social Measurement}, 25, 59-71.
\url{http://doi.org/10.3233/JEM-1999-0160}
}
\keyword{datasets}
