% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{f.aic}
\alias{f.aic}
\title{Compute Akaike information criterion (AIC).}
\usage{
f.aic(theta, y, spec)
}
\arguments{
\item{theta}{Vector (of size d) or matrix (of size M x d) of parameter estimates.}

\item{y}{Vector (of size T) of observations.}

\item{spec}{Model specification created with \code{\link{f.create.spec}}.}
}
\value{
AIC values (scalar or vector of size M).
}
\description{
Compute Akaike information criterion (AIC).
}
\examples{
data("sp500ret")

spec = MSGARCH::f.create.spec(model = c("sGARCH","sGARCH"), distribution = c("norm","norm"),
                             do.skew = c(FALSE,FALSE), do.mix = FALSE, do.shape.ind = FALSE) 

aic = spec$f.aic(theta = spec$theta0, y = sp500ret, spec = spec)
}
\references{
Akaike, H. (1974). A New Look at the Statistical Model Identification. \emph{IEEE Transactions on Automatic Control}, 19, pp. 716-723.
}

