% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{f.bic}
\alias{f.bic}
\title{Compute Bayesian information criterion (BIC).}
\usage{
f.bic(theta, y, spec)
}
\arguments{
\item{theta}{Vector (of size d) or matrix (of size M x d) of parameter estimates.}

\item{y}{Vector (of size T) of observations.}

\item{spec}{Model specification created with \code{\link{f.create.spec}}.}
}
\value{
BIC values (scalar or vector of size M).
}
\description{
Compute Bayesian information criterion (BIC).
}
\examples{
data("sp500ret")

spec = MSGARCH::f.create.spec(model = c("sGARCH","sGARCH"), distribution = c("norm","norm"),
                             do.skew = c(FALSE,FALSE), do.mix = FALSE, do.shape.ind = FALSE) 

bic = spec$f.bic(theta = spec$theta0, y = sp500ret, spec = spec)
}
\references{
Schwarz, G. (1978). Estimating the dimension of a model. \emph{Annals of Statistics}, 6, pp. 461-464.
}

