% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{f.cdf}
\alias{f.cdf}
\title{Cumulative density function at T + 1.}
\usage{
f.cdf(x, theta, y, log = TRUE)
}
\arguments{
\item{x}{Vector (of size N) of point to be evaluated.}

\item{theta}{Vector (of size d) or matrix (of size M x d) of parameter estimates.}

\item{y}{Vector (of size T) of observations.}

\item{log}{Boolean indicating if the log cumulative is returned. (default: \code{log = TRUE})}
}
\value{
Cumulative density or log-density of the points \code{x} (vector of size N or matrix of size M x N).
}
\description{
Function inside a specification returning the cumulative density of a vector of points.
}
\details{
If a matrix of parameter estimates is given, each parameter estimates is evaluated individually.
The \code{f.cdf} function uses the last variance estimate by filtering.
}
\examples{
data("sp500ret")

spec = MSGARCH::f.create.spec(model = c("sGARCH","sGARCH"), distribution = c("norm","norm"),
                             do.skew = c(FALSE,FALSE), do.mix = FALSE, do.shape.ind = FALSE) 
                             
set.seed(123)

x = rnorm(100)

cdf = spec$f.cdf(x = x, theta = spec$theta0, y = sp500ret, log = FALSE)
}

