% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.spec.R
\name{f.create.spec}
\alias{f.create.spec}
\title{Specification creation}
\usage{
f.create.spec(model = c("sGARCH", "sGARCH"), distribution = c("norm",
  "norm"), do.skew = c(FALSE, FALSE), do.mix = FALSE,
  do.shape.ind = FALSE)
}
\arguments{
\item{model}{Vector (of size K) containing the variance model specifications.
 Valid models  are  \code{"sGARCH"},  \code{"eGARCH"},
\code{"gjrGARCH"}, \code{"tGARCH"}, and  \code{"GAS"}. \cr (default: \code{model = c("sGARCH", "sGARCH"}))}

\item{distribution}{Vector (of size K) of conditional densities. Valid
distribution are  \code{"norm"}, \code{"std"}, and  \code{"ged"}. The vector must be of the same length as the models vector. \cr (default: \code{distribution = c("norm", "norm"}))}

\item{do.skew}{Vector (of size K) of boolean indicating if the conditional density is skewed. The vector must be of the same length as the distributions vector. \cr (default: \code{do.skew = c(FALSE, FALSE}))}

\item{do.mix}{Boolean indicating if the specification is a mixture type.  If the argument is \code{TRUE}, a Mixture of GARCH is created,
while if the argument is \code{FALSE}, a Markov-Switching GARCH is created. (default: \code{do.mix = FALSE})}

\item{do.shape.ind}{Boolean indicating if the distribution are Regime-Independent. If the argument is \code{TRUE}, all distributions are
the same and the distribution parameters does not dependent on the regime in which the distribution is attributed to.
If the argument is \code{TRUE}, all distributions in the distribution argument and all skew argument must be the same. (default: \code{do.shape.ind = FALSE})}
}
\value{
A list containing functions and variables related to the created specification. \cr
The list contains:\cr

 Variables:
\itemize{
\item \code{theta0} : Vector (of size d) of default parameters.
\item \code{is.mix} : Boolean indicating if the specification is a mixture.
\item \code{is.shape.ind} : Boolean indicating if the distribution parameter are regime-independent.
\item \code{K} : Number of regimes.
\item \code{sigma0} : Default parameters variance-covariance matrix (of size K x K) used during Bayesian esimation.
\item \code{lower} : Vector (of size d) of lower parameters bound.
\item \code{upper} : Vector (of size d) of upper parameters bound.
\item \code{ineqlb} : Vector (of size d) of lower inequality function bound.
\item \code{inequb} :  Vector (of size d) of upper inequality function bound.
\item \code{n.params} :  Vector (of size K) of the total number of parameters by regime including distribution parameters.
\item \code{n.params.vol} :  Vector (of size K) of the total number of parameters by regime excluding distribuion parameters.
\item \code{do.init} : Boolean indicating the default \code{do.init}  argument.
\item \code{label} : Vector (of size d) of parameters label.
\item \code{name} : Vector (of size K) of specification name.
}
Functions:
\itemize{
\item \code{\link{f.sim}} : Simulation function.
\item \code{\link{f.ht}}  : Conditional variance in each regime.
\item \code{\link{f.kernel}} : Kernel function.
\item \code{\link{f.unc.vol}} : Unconditional variance in each regime.
\item \code{\link{f.pred}} : Predictive density function.
\item \code{\link{f.pit}} : Probability Integral Transform at T + 1.
\item \code{\link{f.risk}} : Value-at-Risk And Expected-Shortfall functions.
\item \code{\link{f.rnd}} : Simulation function at T + 1.
\item \code{\link{f.pdf}} : Probability density function at T + 1.
\item \code{\link{f.cdf}} : Cumulative density function at T + 1.
\item \code{\link{f.Pstate}} : State probabilities filtering function.
\item \code{\link{f.Plast}} : State probabilities at T + 1.
}
}
\description{
Function for creating a variance specification before fitting and using the \R package \code{MSGARCH} functionalities.
}
\details{
The Markov-Switching specification created is based on the Haas et al. (2004a) MSGARCH specification.
  It is a MSGARCH model that is separated in K single-regimes specification  which are updated in parallel.
  Under this specification, the conditional variance is a function of the past data and the current state.
  The Mixture of GARCH option is based on the Haas et al. (2004b). A Mixture of GARCH is a mixture of distribution 
  where the variance process of each distribution is a single-regime process.
}
\examples{

spec = MSGARCH::f.create.spec(model = c("sGARCH","gjrGARCH"), distribution = c("norm","std"),
                             do.skew = c(TRUE,FALSE), do.mix = FALSE, do.shape.ind = FALSE) 
}
\references{
Bollerslev, T. (1986). Generalized Autoregressive Conditional Heteroskedasticity. \emph{Journal of Econometrics}, 31, pp. 307-327.

Creal, D. Koopman, S. J. & Lucas, A. (2013). Generalized Autoregressive Score Models with Applications. \emph{Journal of Applied Econometrics}, 28, pp. 777-795.

Fernández, C. & Steel, M. F. (1998). On Bayesian Modeling of Fat Tails and Skewness. \emph{Journal of the American Statistical Association}, 93, pp. 359-371.

Glosten, L. R. Jagannathan, R. & Runkle, D. E. (1993). On the Relation Between the Expected Value and the Volatility of the Nominal Excess Return on Stocks. \emph{Journal of Finance}, 48, pp. 1779-1801.

Haas, M. Mittnik, S. & Paolella, M. S. (2004a). A New Approach to Markov-Switching GARCH Models. \emph{Journal of Financial Econometrics}, 2, pp. 493-530.

Haas, M. Mittnik, S. & Paolella, M. S. (2004b). Mixed Normal Conditional Heteroskedasticity. \emph{Journal of Financial Econometrics}, 2, pp. 211-250.

Nelson, D. B. (1991). Conditional Heteroskedasticity in Asset Returns: A New Approach. \emph{Econometrica}, 59, pp. 347-370.

Zakoian, J.-M. (1994). Threshold Heteroskedastic Models Journal of Economic. \emph{Dynamics and Control}, 18, pp. 931-955.
}

