% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{f.estim.mle}
\alias{f.estim.mle}
\title{ML estimation.}
\usage{
f.estim.mle(y, spec, ctr = list())
}
\arguments{
\item{y}{Vector (of size T) of observations.}

\item{spec}{Model specification created with \code{\link{f.create.spec}}.}

\item{ctr}{List of control parameters.
The control parameters have two components to it:
\itemize{
\item \code{theta0} : Starting parameters (vector of size d). If no starting parameters is provided, the default starting parameters of the specification are used.
\item \code{do.init} : Boolean indicating if there is a pre-optimization with the \R package \code{DEoptim} (Ardia et al., 2011). (default: \code{do.init = FALSE})
}}
}
\value{
A list containing two variables:
       \itemize{
       \item \code{theta} : Optimal parameters (vector of size d).
       \item \code{nll_likelihood} : Negative log-likelihood of y given the optimal parameters.
       }
}
\description{
Function that performs Maximum Likelihood estimation of a \code{\link{MSGARCH}} specification on a set of observations.
}
\details{
The Maximum likelihood estimation uses the \R package \code{nloptr} (Johnson, 2014) for main optimizer 
while it uses the \R package \code{DEoptim} when \code{do.init = TRUE}.
}
\examples{
data("sp500ret")

spec = MSGARCH::f.create.spec(model = c("sGARCH","sGARCH"), distribution = c("norm","norm"),
                              do.skew = c(FALSE,FALSE), do.mix = FALSE, do.shape.ind = FALSE) 

theta = MSGARCH::f.estim.mle(y = sp500ret, spec = spec,  ctr = list(do.init = TRUE))
}
\references{
Ardia, D.; Mullen, K. M.; Peterson, B. G. & Ulrich, J. (2015). \code{DEoptim}: Differential Evolution in \R. \url{https://cran.r-project.org/web/packages/DEoptim/}.

Johnson, S. G. (2014). The NLopt Nonlinear-Optimization. \url{https://cran.r-project.org/web/packages/NLopt/}.
}

