% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{f.ht}
\alias{f.ht}
\title{Conditional variance in each regime.}
\usage{
f.ht(theta, y)
}
\arguments{
\item{theta}{Vector (of size d) or matrix (of size M x d) of parameter estimates.}

\item{y}{Vector (of size T) of observations.}
}
\value{
Condititional variance time serie (array of size T + 1 x M x K) for each regime.
}
\description{
Function inside a specification returning the conditional variance of each regime.
}
\details{
If a matrix of parameter estimates is given, each parameter estimates is evaluated individually.
}
\examples{
data("sp500ret")

spec = MSGARCH::f.create.spec(model = c("sGARCH","sGARCH"), distribution = c("norm","norm"),
                             do.skew = c(FALSE,FALSE), do.mix = FALSE, do.shape.ind = FALSE) 

ht = spec$f.ht(theta = spec$theta0, y = sp500ret)
}

