% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{f.pit}
\alias{f.pit}
\title{Probability Integral Transform at T + 1.}
\usage{
f.pit(x, theta, y, spec, do.norm = FALSE)
}
\arguments{
\item{x}{Vector (of size N) of point to be evaluated}

\item{theta}{Vector (of size d) or matrix (of size M x d) of parameter estimates.}

\item{y}{Vector (of size T) of observations.}

\item{do.norm}{Boolean indicating if the PIT value are transforms into standard Normal variate. (\code{do.norm = FALSE})}
}
\value{
Probability integral transform of the points \code{x} or Normal variate derived from the Probability integral transform of \code{x} (vector of size N).
}
\description{
Function inside a specification returning the predictive Probability integral transform (PIT).
}
\details{
If a matrix of MCMC posterior draws estimates is given, the Bayesian Probability integral transform is calculated.
         The \code{do.norm} argument transforms the PIT value into Normal variate so that normality test can be done.
}
\examples{
data("sp500ret")

spec = MSGARCH::f.create.spec(model = c("sGARCH","sGARCH"), distribution = c("norm","norm"),
                             do.skew = c(FALSE,FALSE), do.mix = FALSE, do.shape.ind = FALSE) 
                             
set.seed(123)

x = rnorm(100)

pit = spec$f.pit(x = x theta = spec$theta0, y = sp500ret, do.norm = FALSE)
}

