% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{f.Plast}
\alias{f.Plast}
\title{State probabilities at T + 1.}
\usage{
f.Plast(theta, y)
}
\arguments{
\item{theta}{Vector (of size d) or matrix (of size M x d) of parameter estimates.}

\item{y}{Vector (of size T) of observations.}
}
\value{
State probabilities at T + 1 (matrix of size M x K).
}
\description{
Function inside a specification returning the state probabilities at  T + 1.
}
\details{
If a matrix of parameter estimates is given, each parameter estimates is evaluated individually.
}
\examples{
data("sp500ret")

spec = MSGARCH::f.create.spec(model = c("sGARCH","sGARCH"), distribution = c("norm","norm"),
                             do.skew = c(FALSE,FALSE), do.mix = FALSE, do.shape.ind = FALSE) 

Plast = spec$f.Plast(theta = spec$theta0, y = sp500ret) 
}

