% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{f.pred}
\alias{f.pred}
\title{Predictive density function.}
\usage{
f.pred(x, theta, y, log = TRUE)
}
\arguments{
\item{x}{Vector (of size N) of point to be evaluated}

\item{theta}{Vector (of size d) or matrix (of size M x d) of parameter estimates.}

\item{y}{Vector (of size T) of observations.}

\item{log}{Boolean indicating if the log-density is returned. (default: \code{log = TRUE})}
}
\value{
Predictive density or log-density of \code{x} (vector of size N).
}
\description{
Function inside a specification returning the predictive probability density of a vector of points.
}
\details{
If a matrix of MCMC posterior draws estimates is given, the Bayesian predictive density is calculated.
}
\examples{
data("sp500ret")

spec = MSGARCH::f.create.spec(model = c("sGARCH","sGARCH"), distribution = c("norm","norm"),
                             do.skew = c(FALSE,FALSE), do.mix = FALSE, do.shape.ind = FALSE) 
                             
set.seed(123)

x = rnorm(100)

pred = spec$f.pred(x = x, theta = spec$theta0, y = sp500ret, log = TRUE)
}

