% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{f.Pstate}
\alias{f.Pstate}
\title{State probabilities filtering function.}
\usage{
f.Pstate(theta, y)
}
\arguments{
\item{theta}{Vector (of size d) or matrix (of size M x d) of parameter estimates.}

\item{y}{Vector (of size T) of observations.}
}
\value{
Filtered state probabilities (array of size T x M x K).
}
\description{
Function inside a specification returning the filtered state probabilities.
}
\details{
If a matrix of parameter estimates is given, each parameter estimates is evaluated individually.
}
\examples{
data("sp500ret")

spec = MSGARCH::f.create.spec(model = c("sGARCH","sGARCH"), distribution = c("norm","norm"),
                             do.skew = c(FALSE,FALSE), do.mix = FALSE, do.shape.ind = FALSE) 

Pstate  = spec$f.Pstate(theta = spec$theta0, y = sp500ret)
}

