% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{f.risk}
\alias{f.risk}
\title{Value-at-Risk And Expected-shortfall functions.}
\usage{
f.risk(theta, y, level = c(0.95, 0.99), ES = TRUE)
}
\arguments{
\item{theta}{Vector (of size d) or matrix (of size M x d) of parameter estimates.}

\item{y}{Vector (of size T) of observations.}

\item{level}{Vector (of size A) of Value-at-risk and Expected-shortfall levels. (default: \code{level = c(0.95,0.99)})}

\item{ES}{Boolean indicating if Expected-shortfall is also calculated. (default: \code{ES = TRUE})}
}
\value{
A list containing one or two components:
\itemize{
\item VaR : Value-at-Risk at the choosen levels (vector of size A).
\item ES  : Expected-shortfall at the choosen levels (vector of size A).
}
}
\description{
Function inside a specification returning the Value-at-risk and Expected-shortfall.
}
\details{
If a matrix of MCMC posterior draws estimates is given, the Bayesian Value-at-Risk and Expected-shortfall are calculated.
}
\examples{
data("sp500ret")

spec = MSGARCH::f.create.spec(model = c("sGARCH","sGARCH"), distribution = c("norm","norm"),
                             do.skew = c(FALSE,FALSE), do.mix = FALSE, do.shape.ind = FALSE) 

risk = spec$f.risk(theta = spec$theta0, y = sp500ret,level = c(0.95,0.99), ES = TRUE)
}

