% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{f.rnd}
\alias{f.rnd}
\title{Simulation function at T + 1.}
\usage{
f.rnd(n, theta, y, do.state = FALSE)
}
\arguments{
\item{n}{Number of random draws to be generated.}

\item{theta}{Vector (of size d) or matrix (of size M x d) of parameter estimates.}

\item{y}{Vector (of size T) of observations.}

\item{do.state}{Boolean indicating if the simulated state are also output. (default: \code{do.state = FALSE})}
}
\value{
A list containing one or two components:
\itemize{
\item \code{draws}: vector (of size n) or matrix (of size M x n) of simulated draws at T + 1.
\item \code{state}: vector (of size n) or matrix (of size M x n) of simulated states at T + 1.
 The \code{state} value appear only if \code{do.state = TRUE}.
}
}
\description{
Function inside a specification returning random draws at T + 1.
}
\details{
If a matrix of parameter estimates is given, each parameter estimates is evaluated individually.
}
\examples{
spec = MSGARCH::f.create.spec(model = c("sGARCH","sGARCH"), distribution = c("norm","norm"),
do.skew = c(FALSE,FALSE), do.mix = FALSE, do.shape.ind = FALSE) 

rnd = spec$f.rnd(n = 1000, theta = theta, y = sp500ret, do.state = TRUE)
}

