% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{f.sim}
\alias{f.sim}
\title{Simulation function.}
\usage{
f.sim(n, theta, burnin = 500, do.state = FALSE)
}
\arguments{
\item{n}{Simulation length.}

\item{theta}{Vector (of size d) or matrix (of size M x d) of parameter estimates.}

\item{burnin}{(integer >= 0) Burnin period discarded (first simulation draws). (default: \code{burnin = 500})}

\item{do.state}{Boolean  indicating if the simulated state are also output. (default: \code{log = TRUE})}
}
\value{
A list containing one or two components.
\itemize{
\item \code{draws}: vector (of size n) or matrix (of size M x n) of simulated draws.
\item \code{state}: vector (of size n) or matrix (of size M x n) of simulated states.
 The \code{state} value appear only if \code{do.state = TRUE}.
}
}
\description{
Function inside a specification returning a simulated process.
}
\details{
If a matrix of parameter estimates is given, each parameter estimates is evaluated individually.
}
\examples{
spec = MSGARCH::f.create.spec(model = c("sGARCH","sGARCH"), distribution = c("norm","norm"),
do.skew = c(FALSE,FALSE), do.mix = FALSE, do.shape.ind = FALSE) 

y = spec$f.sim(n = 1000, theta = spec$theta0, burnin = 500, do.state = TRUE)

}

