% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{f.unc.vol}
\alias{f.unc.vol}
\title{Unconditional variance of each regime.}
\usage{
f.unc.vol(theta)
}
\arguments{
\item{theta}{Vector (of size d) or matrix (of size M x d) of parameter estimates.}
}
\value{
Unconditional variance (vector of size K or matrix of size M x K) of each regime.
}
\description{
Function inside a specification returning the unconditional variance of the process in each state.
}
\details{
If a matrix of parameter estimates is given, each parameter estimates is evaluated individually.
}
\examples{
spec = MSGARCH::f.create.spec(model = c("sGARCH","sGARCH"), distribution = c("norm","norm"),
                             do.skew = c(FALSE,FALSE), do.mix = FALSE, do.shape.ind = FALSE) 

unc.vol = spec$f.unc.vol(theta = spec$theta0)
}

