% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht.R
\name{ht}
\alias{ht}
\title{Conditional volatility in each regime.}
\usage{
ht(spec, theta, y)
}
\arguments{
\item{spec}{Model specification of class \code{MSGARCH_SPEC} created with \code{\link{create.spec}}.}

\item{theta}{Vector (of size d) or matrix (of size M x d) of parameter estimates.}

\item{y}{Vector (of size T) of observations.}
}
\value{
Condititional volatility time serie (array of size (T + 1) x M x K) in each regime.
}
\description{
Method returning the conditional volatility in each regime.
}
\details{
If a matrix of parameter estimates is given, each parameter estimates is evaluated individually.
}
\examples{
# load data
data("sp500ret")

# create model specification
spec = MSGARCH::create.spec(model = c("sGARCH","sGARCH"), distribution = c("norm","norm"),
                             do.skew = c(FALSE,FALSE), do.mix = FALSE, do.shape.ind = FALSE) 

# Compute the conditional volatility
ht = MSGARCH::ht(spec = spec,theta = spec$theta0, y = sp500ret)

plot(ht)
}

