% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plast.R
\name{Plast}
\alias{Plast}
\title{State probabilities at T + 1.}
\usage{
Plast(spec, theta, y)
}
\arguments{
\item{spec}{Model specification of class \code{\link{MSGARCH_SPEC}} created with \code{\link{create.spec}}.}

\item{theta}{Vector (of size d) or matrix (of size M x d) of parameter estimates.}

\item{y}{Vector (of size T) of observations.}
}
\value{
State probabilities at T + 1 (matrix of size M x K).
}
\description{
Method returning the state probabilities at  T + 1.
}
\details{
If a matrix of parameter estimates is given, each parameter estimates is evaluated individually.
}
\examples{
data("sp500ret")

spec = MSGARCH::create.spec(model = c("sGARCH","sGARCH"), distribution = c("norm","norm"),
                             do.skew = c(FALSE,FALSE), do.mix = FALSE, do.shape.ind = FALSE) 

Plast = MSGARCH::Plast(spec = spec, theta = theta, y = sp500ret) 
}

