% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred}
\alias{pred}
\title{Predictive function.}
\usage{
pred(object, x, theta, y, log = FALSE, do.its = FALSE)
}
\arguments{
\item{object}{Model specification of class \code{MSGARCH_SPEC} created with \code{\link{create.spec}}
or fit object of type \code{MSGARCH_MLE_FIT} created with \code{\link{fit.mle}} or \code{MSGARCH_BAY_FIT}
created with \code{\link{fit.bayes}}.}

\item{x}{Vector (of size N).}

\item{theta}{Vector (of size d) or matrix (of size M x d) of parameter estimates (not required when using a fit object) where d must have
the same length as the default parameters of the specification.}

\item{y}{Vector (of size T) of observations (not required when using a fit object).}

\item{log}{Boolean indicating if the log-density is returned. (Default: \code{log = FALSE})}

\item{do.its}{Boolean indicating if the in-sample predictive is returned. (Default: \code{do.its = FALSE})}
}
\value{
A list of class \code{MSGARCH_PRED} containing two components:
\itemize{
\item \code{pred}:\cr If \code{do.its = FALSE}: (Log-)Predictive of of the points \code{x} at \code{t = T + 1} (vector of size N). \cr
                  If \code{do.its = TRUE}: In-sample Predictive of \code{y} (vector of size T or matrix of size M x T). 
\item \code{do.its}: Original user inputed \code{do.its} for reference.
}
}
\description{
Method returning the predictive probability density in-sample or of a vector of points  consider as one step ahead draws (\code{t = T + 1}).
}
\details{
If a matrix of parameter estimates is given, each parameter estimate (each row) is evaluated individually. 
If \code{do.its = FALSE}, the vector \code{x} are evaluated as \code{t = T + 1} realization and the method uses the variance estimate at \code{t = T + 1}.
If \code{do.its = TRUE} and  each column of  \code{x} is evaluated a their respective time \code{t} indicated by their column index.
Finally if \code{x = NULL} the vector \code{y} is evaluated using their respective variance estimate at each time \code{t}.
}
\examples{
require("MSGARCH")
# load data
data("sp500")
sp500 = sp500[1:1000]

# create model specification
spec = MSGARCH::create.spec() 

# fit the model on the data with ML estimation using DEoptim intialization
set.seed(123)
fit = MSGARCH::fit.mle(spec = spec, y = sp500, ctr = list(do.init = FALSE))
                           
# run pred method in-sample     
pred.its = MSGARCH::pred(object = fit, log = TRUE, do.its = TRUE)  

sum(pred.its$pred, na.rm = TRUE)
                                             
# create mesh
x = seq(-3,3,0.01)

# run pred method on mesh at T + 1
pred = MSGARCH::pred(object = fit, x = x, log = TRUE, do.its = FALSE)

plot(pred)
}
