% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk}
\alias{risk}
\title{Value-at-Risk And Expected-shortfall functions at T + 1.}
\usage{
risk(spec, theta, y, level = c(0.95, 0.99), ES = TRUE)
}
\arguments{
\item{spec}{Model specification of class \code{MSGARCH_SPEC} created with \code{\link{create.spec}}.}

\item{theta}{Vector (of size d) or matrix (of size M x d) of parameter estimates.}

\item{y}{Vector (of size T) of observations.}

\item{level}{Vector (of size R) of Value-at-risk and Expected-shortfall levels.\cr
(Default: \code{level = c(0.95,0.99)})}

\item{ES}{Boolean indicating if Expected-shortfall is also calculated. (Default: \code{ES = TRUE})}
}
\value{
A list containing one or two components:
\itemize{
\item \code{VaR} : Value-at-Risk at T + 1 at the choosen levels (vector of size R).
\item \code{ES}  : Expected-shortfall at T + 1 at the choosen levels (vector of size R).
}
}
\description{
Method returning the Value-at-Risk and Expected-shortfall at T + 1.
}
\details{
If a matrix of MCMC posterior draws estimates is given, the Bayesian Value-at-Risk and Expected-shortfall are calculated.
}
\examples{
# load data
data("sp500ret")

# create model specification
spec = MSGARCH::create.spec(model = c("sGARCH","sGARCH"), distribution = c("norm","norm"),
                             do.skew = c(FALSE,FALSE), do.mix = FALSE, do.shape.ind = FALSE) 
set.seed(123)
                             
# compute the Value-at-Risk and Expected-shortfall 
risk = MSGARCH::risk(spec = spec, theta = spec$theta0, y = sp500ret,
                    level = c(0.95,0.99), ES = TRUE)
}

