% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{Process simulation method.}
\usage{
sim(spec, n, m, theta, burnin = 500)
}
\arguments{
\item{spec}{Model specification of class \code{MSGARCH_SPEC} created with \code{\link{create.spec}}.}

\item{n}{Simulation length.}

\item{m}{Number of simulation.}

\item{theta}{Vector (of size d) or matrix (of size M x d) of parameter estimates.}

\item{burnin}{(integer >= 0) Burnin period discarded (first simulation draws). (Default: \code{burnin = 500})}
}
\value{
A list of class \code{MSGARCH_SIM} containing two components.
\itemize{
\item \code{draws}: Matrix (of size M x n) of simulated draws.
\item \code{state}: Matrix (of size M x n) of simulated states.
}
 The \code{MSGARCH_SIM} class contains the \code{plot} method.
}
\description{
Method returning a \code{MSGARCH} specification process.
}
\details{
If a matrix of parameter estimates is given, each parameter estimates is evaluated individually and \code{m = nrow(theta)}.
}
\examples{
# create model specification
spec = MSGARCH::create.spec() 

# generate process
set.seed(123)
sim = MSGARCH::sim(spec = spec, n = 1000, m = 1, theta = spec$theta0, burnin = 500)

plot(sim)
}

