% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSGARCH.R
\docType{data}
\name{SMI}
\alias{SMI}
\title{Swiss market index dataset}
\format{
\code{zoo} object containing 2,500 observations ranging from 1990-11-12 to 2000-10-20.
}
\source{
\code{DEoptim} package
}
\usage{
data("SMI")
}
\description{
See Mullen et al. (2011) for a description of this dataset.
}
\references{
Mullen, K.M, Ardia, D., Gil, D., Windover, D., Cline, J. (2011).
\code{DEoptim}: An R Package for Global Optimization by Differential Evolution. 
\emph{Journal of Statistical Software}, 40(6), 1-26.
\doi{10.18637/jss.v040.i06}
}
\keyword{datasets}
