\name{sp500}
\docType{data}
\alias{sp500}
\title{Log return of the S&P 500 index closing Value}
\description{
The S&P 500 index closing value log return from 1998-01-01 to 2015-12-31 from
Yahoo Finance \url{https://finance.yahoo.com/}.
}
\usage{data("sp500")}
\format{Matrix containing 4529 observations.}
\source{Yahoo Finance \url{https://finance.yahoo.com/}}
\keyword{datasets}
