\name{sp500ret}
\docType{data}
\alias{sp500ret}
\title{Log return of the S&P 500 index closing Value}
\description{
The S&P 500 index closing value log return from 1987-03-10 to 2009-01-30 from
Yahoo Finance \url{https://ca.finance.yahoo.com/}.
}
\usage{data("sp500ret")}
\format{Vector containing 5,523 observations.}
\source{Yahoo Finance \url{https://ca.finance.yahoo.com/}}
\keyword{datasets}
