% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.R
\name{AIC}
\alias{AIC}
\title{Compute Akaike information criterion (AIC).}
\usage{
AIC(fit)
}
\arguments{
\item{fit}{Fit object of type \code{MSGARCH_MLE_FIT} created with \code{\link{fit.mle}} or \code{MSGARCH_BAY_FIT} created with \code{\link{fit.bayes}}.}
}
\value{
AIC value.
}
\description{
Compute Akaike information criterion (AIC).
}
\details{
If a matrix of MCMC posterior draws estimates is given, the AIC on the posterior mean is calculated.
}
\examples{
# load data
data("sp500ret")

# create model specification
spec = MSGARCH::create.spec(model = c("sGARCH","sGARCH"), distribution = c("norm","norm"),
                             do.skew = c(FALSE,FALSE), do.mix = FALSE, do.shape.ind = FALSE) 

set.seed(123)

# fit the model by MLE                                                         
fit = MSGARCH::fit.mle(spec = spec, y = sp500ret)

# compute AIC
AIC = MSGARCH::AIC(fit = fit)
}
\references{
Akaike, H. (1974). A New Look at the Statistical Model Identification. \emph{IEEE Transactions on Automatic Control}, 19, pp. 716-723.
}

