% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIC.R
\name{DIC}
\alias{DIC}
\title{Compute Deviance Information Criterion (DIC).}
\usage{
DIC(fit)
}
\arguments{
\item{fit}{Fit object of type \code{\link{MSGARCH_BAY_FIT}} created with \code{\link{fit.bayes}}.}
}
\value{
A list containing six variables:
       \itemize{
       \item \code{DIC} : Deviance Information Criterion.
       \item \code{IC}  : Bayesian Predictive Information Criterion.
       \item \code{pD}  : Effective number of parameters (pD = Dbar - Dhat)
       \item \code{pV}  : Effective number of parameters (pV = var(D)/2)
       \item \code{D.bar}: Expected value of the deviance over the posterior
       \item \code{D.hat}: Deviance at the mean posterior estimate
       }
}
\description{
Compute Deviance Information Criterion (DIC).
}
\examples{
data("sp500ret")

spec = MSGARCH::create.spec(model = c("sGARCH","sGARCH"), distribution = c("norm","norm"),
                             do.skew = c(FALSE,FALSE), do.mix = FALSE, do.shape.ind = FALSE) 
                             
fit = MSGARCH::fit.bay(spec = spec, y = sp500ret)

DIC = MSGARCH::DIC(fit = fit)
}
\references{
Gelman, A. Carlin, J. B. Stern, H. S. & Rubin, D. B. (2003). Bayesian Data Analysis. \emph{Chapman and Hall/CRC}
}

