% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIC.R
\name{DIC}
\alias{DIC}
\title{Compute Deviance Information Criterion (DIC).}
\usage{
DIC(fit)
}
\arguments{
\item{fit}{Fit object of type \code{MSGARCH_BAY_FIT} created with \code{\link{fit.bayes}}.}
}
\value{
A list containing six variables:
       \itemize{
       \item \code{DIC} : Deviance Information Criterion.
       \item \code{IC}  : Bayesian Predictive Information Criterion.
       \item \code{pD}  : Effective number of parameters (pD = Dbar - Dhat)
       \item \code{pV}  : Effective number of parameters (pV = var(D)/2)
       \item \code{D.bar}: Expected value of the deviance over the posterior
       \item \code{D.hat}: Deviance at the mean posterior estimate
       }
}
\description{
Compute Deviance Information Criterion (DIC).
}
\examples{
# load data
data("sp500ret")

# create model specification
spec = MSGARCH::create.spec(model = c("sGARCH","sGARCH"), distribution = c("norm","norm"),
                             do.skew = c(FALSE,FALSE), do.mix = FALSE, do.shape.ind = FALSE) 

set.seed(123)

# fit the model by Bayesian estimation                                                            
fit = MSGARCH::fit.bayes(spec = spec, y = sp500ret)

# compute DIC
DIC = MSGARCH::DIC(fit = fit)
}
\references{
Gelman, A. Carlin, J. B. Stern, H. S. & Rubin, D. B. (2003). Bayesian Data Analysis. \emph{Chapman and Hall/CRC}
}

