% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{fit.mle}
\alias{fit.mle}
\title{ML estimation.}
\usage{
fit.mle(spec, y, ctr = list())
}
\arguments{
\item{spec}{Model specification created with \code{\link{create.spec}}.}

\item{y}{Vector (of size T) of observations.}

\item{ctr}{List of control parameters.
The control parameters have two components to it:
\itemize{
\item \code{theta0} : Starting parameters (vector of size d). If no starting parameters is provided, the default starting parameters of the specification are used.
\item \code{do.init} : Boolean indicating if there is a pre-optimization with the \R package \code{DEoptim} (Ardia et al., 2011). (Default: \code{do.init = FALSE})
\item \code{NP} : Number of parameter vectors in the population in \code{DEoptim} optimization. (Default: \code{NP = 500})
\item \code{itermax} : Maximum iteration (population generation) allowed in \code{DEoptim} optimization. (Default: \code{maxit = 500})
}}
}
\value{
A list of class \code{MSGARCH_MLE_FIT} containing five components:
       \itemize{
       \item \code{theta} : Optimal parameters (vector of size d).
       \item \code{ll_likelihood} : log-likelihood of \code{y} given the optimal parameters.
       \item \code{spec} : Specification.
       \item \code{is.init} : Indicating if estimation was made with do.init option.
       \item \code{y} :  Vector (of size T) of observations..
       }
The \code{MSGARCH_MLE_FIT} contains these methods:
\itemize{
\item \code{\link{AIC}} : Compute Akaike information criterion (AIC).
\item \code{\link{BIC}} : Compute Bayesian information criterion (BIC).
}
}
\description{
Method that performs Maximum Likelihood estimation of a \code{MSGARCH_SPEC} object on a set of observations.
}
\details{
The Maximum likelihood estimation uses the \R package \code{nloptr} (Johnson, 2014) for main optimizer 
while it uses the \R package \code{DEoptim} when \code{do.init = TRUE}.
}
\examples{
# load data
data("sp500ret")


# create model specification
spec = MSGARCH::create.spec(model = c("sGARCH","sGARCH"), distribution = c("norm","norm"),
                              do.skew = c(FALSE,FALSE), do.mix = FALSE, do.shape.ind = FALSE) 

set.seed(123)

# fit the model on the data with ML estimation
fit = MSGARCH::fit.mle(spec = spec, y = sp500ret, 
                       ctr = list(do.init = TRUE, NP = 100, itermax = 100))
}
\references{
Ardia, D.; Mullen, K. M.; Peterson, B. G. & Ulrich, J. (2015). \code{DEoptim}: Differential Evolution in \R. \url{https://cran.r-project.org/web/packages/DEoptim/}.

Johnson, S. G. (2014). The NLopt Nonlinear-Optimization. \url{https://cran.r-project.org/web/packages/NLopt/}.
}

