% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht.R
\name{ht}
\alias{ht}
\title{Conditional volatility in each regime.}
\usage{
ht(spec, theta, y)
ht(fit)
}
\arguments{
\item{spec}{Model specification of class \code{MSGARCH_SPEC} created with \code{\link{create.spec}}.}

\item{theta}{Vector (of size d) or matrix (of size M x d) of parameter estimates.}

\item{y}{Vector (of size T) of observations.}

\item{fit}{Fit object of type \code{MSGARCH_MLE_FIT} created with \code{\link{fit.mle}} or \code{MSGARCH_BAY_FIT} created with \code{\link{fit.bayes}}.}
}
\value{
Condititional volatility time serie (array of size (T + 1) x M x K) in each regime.
}
\description{
Method returning the conditional volatility in each regime.
}
\details{
If a matrix of parameter estimates is given, each parameter estimates is evaluated individually.
}
\examples{
\dontrun{
# load data
data("sp500ret")

# create model specification
spec = MSGARCH::create.spec() 

# fit the model on the data with ML estimation using DEoptim intialization
set.seed(123)
fit = MSGARCH::fit.mle(spec = spec, y = sp500ret)

# Compute the conditional volatility
ht = MSGARCH::ht(fit)

plot(ht)
}
}

