% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSGARCH.R
\docType{package}
\name{MSGARCH-package}
\alias{MSGARCH}
\alias{MSGARCH-package}
\title{The R package MSGARCH}
\description{
The \R package \pkg{MSGARCH} implements a comprehensive
set of functionalities for Markov-switching GARCH (Haas et al. 2004a) and Mixture of GARCH (Haas et al. 2004b) models,
This includes fitting, filtering, forecasting, and simulating.
Other functions related to Value-at-Risk and Expected-Shortfall are also available.\cr
The main functions of the package are coded
in \code{C++} using \pkg{Rcpp} (Eddelbuettel and Francois, 2011)
and \pkg{RcppArmadillo} (Eddelbuettel and Sanderson, 2014).\cr
\pkg{MSGARCH} focuses on the conditional variance (and higher moments) process.
Hence, there is no equation for the mean.
Therefore, you must pre-filter via AR(1) before applying the model.\cr
The \pkg{MSGARCH} package implements a variety of GARCH specifications together with several conditional distributions.
This allows for a rich modeling
environment for Markov-switching GARCH models. Each single-regime process
is a one-lag process (e.g., GARCH(1,1)).
When optimization is performed, we ensure that the variance in each regime is covariance-stationary
and strictly positive (refer to the vignette for more information).\cr
We refer to Ardia et al. (2019a) for a detailed
introduction to the package and its usage. Refer to Ardia et al. (2018) and Ardia et al. (2019b) for
further applications. \cr
The authors acknowledge Google for financial support via the Google Summer of Code 2016 & 2017,
the International Institute of Forecasters and Industrielle-Alliance.
}
\note{
By using \pkg{MSGARCH} you agree to the following rules:
\itemize{
\item You must cite Ardia et al. (2019a) in working papers and published papers that use \pkg{MSGARCH}.
Use \code{citation("MSGARCH")}.
\item You must place the following URL in a footnote to help others find \pkg{MSGARCH}: 
\url{https://CRAN.R-project.org/package=MSGARCH}.
\item You assume all risk for the use of \pkg{MSGARCH}.
}
}
\references{
Ardia, D. Bluteau, K. Boudt, K. Catania, L. (2018).
Forecasting risk with Markov-switching GARCH models: A large-scale performance study.
\emph{International Journal of Forecasting}, 34(4), 733-747.
\url{http://doi.org/10.1016/j.ijforecast.2018.05.004}

Ardia, D. Bluteau, K. Boudt, K. Catania, L. Trottier, D.-A. (2019a).
Markov-switching GARCH models in \R: The \pkg{MSGARCH} package.
\emph{Journal of Statistical Software}, 91(4), 1-38.
\url{http://doi.org/10.18637/jss.v091.i04}

Ardia, D. Bluteau, K. Ruede, M. (2019b).
Regime changes in Bitcoin GARCH volatility dynamics.
\emph{Finance Research Letters}, 29, 266-271.
\url{http://doi.org/10.1016/j.frl.2018.08.009}

Eddelbuettel, D. & Francois, R. (2011).
\pkg{Rcpp}: Seamless \R and \code{C++} integration.
\emph{Journal of Statistical Software}, 40, 1-18.
\url{http://doi.org/10.18637/jss.v040.i08}

Eddelbuettel, D. & Sanderson, C. (2014).
\pkg{RcppArmadillo}: Accelerating \R with high-performance \code{C++} linear algebra.
\emph{Computational Statistics & Data Analysis}, 71, 1054-1063.
\url{http://doi.org/10.1016/j.csda.2013.02.005}

Haas, M. Mittnik, S. & Paolella, MS. (2004).
A new approach to Markov-switching GARCH models.
\emph{Journal of Financial Econometrics}, 2, 493-530.
\url{http://doi.org/10.1093/jjfinec/nbh020}

Haas, M. Mittnik, S. & Paolella, M. S. (2004b).
Mixed normal conditional heteroskedasticity.
\emph{Journal of Financial Econometrics}, 2, 211-250.
\url{http://doi.org/10.1093/jjfinec/nbh009}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/keblu/MSGARCH}
  \item Report bugs at \url{https://github.com/keblu/MSGARCH/issues}
}

}
\author{
\strong{Maintainer}: Keven Bluteau \email{Keven.Bluteau@hec.ca}

Authors:
\itemize{
  \item David Ardia \email{david.ardia.ch@gmail.com}
  \item Leopoldo Catania \email{leopoldo.catania@econ.au.dk}
  \item Denis-Alexandre Trottier \email{denis-alexandre.trottier.1@ulaval.ca}
}

Other contributors:
\itemize{
  \item Kris Boudt \email{kris.boudt@vub.ac.be} [contributor]
  \item Alexios Ghalanos \email{alexios@4dscape.com} [contributor]
  \item Brian Peterson \email{brian@braverock.com} [contributor]
}

}
