% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{pdf}
\alias{pdf}
\title{Probability density function at T + 1.}
\usage{
pdf(spec, x, theta, y, log = TRUE)
}
\arguments{
\item{spec}{Model specification of class \code{MSGARCH_SPEC} created with \code{\link{create.spec}}.}

\item{x}{Vector (of size N) of point to be evaluated}

\item{theta}{Vector (of size d) or matrix (of size M x d) of parameter estimates.}

\item{y}{Vector (of size T) of observations.}

\item{log}{Boolean indicating if the log-density is returned. (default: \code{log = TRUE})}
}
\value{
Probability density or log-density of the points \code{x} (vector of size N or matrix of size M x N).
}
\description{
Method returning the probability density of a vector of points.
}
\details{
If a matrix of parameter estimates is given, each parameter estimates is evaluated individually.
 The \code{\link{pdf}} method uses the last variance estimate by filtering.
}
\examples{
data("sp500ret")

spec = MSGARCH::create.spec(model = c("sGARCH","sGARCH"), distribution = c("norm","norm"),
                           do.skew = c(FALSE,FALSE), do.mix = FALSE, do.shape.ind = FALSE) 
                             
set.seed(123)

x = rnorm(100)

pdf = MSGARCH::pdf(spec = spec, x = x, theta = spec$theta, y = sp500ret, log = FALSE)
}

