% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pstate.R
\name{Pstate}
\alias{Pstate}
\title{Filtered state probabilities.}
\usage{
Pstate(spec, theta, y)
Pstate(fit)
}
\arguments{
\item{spec}{Model specification of class \code{MSGARCH_SPEC} created with \code{\link{create.spec}}.}

\item{theta}{Vector (of size d) or matrix (of size M x d) of parameter estimates.}

\item{y}{Vector (of size T) of observations.}

\item{fit}{Fit object of type \code{MSGARCH_MLE_FIT} created with \code{\link{fit.mle}} or \code{MSGARCH_BAY_FIT} created with \code{\link{fit.bayes}}.}
}
\value{
Filtered state probabilities of class \code{MSGARCH_RND} (array of size (T + 1) x M x K).
The class \code{MSGARCH_RND} contains the \code{plot} method.
}
\description{
Method returning the filtered state probabilities.
}
\details{
If a matrix of parameter estimates is given, each parameter estimates is evaluated individually.
}
\examples{
\dontrun{
# load data
data("sp500ret")

# create model specification
spec = MSGARCH::create.spec() 

# fit the model on the data with ML estimation using DEoptim intialization
set.seed(123)
fit = MSGARCH::fit.mle(spec = spec, y = sp500ret)
 
# compute the filtered state probabilities
Pstate  = MSGARCH::Pstate(fit)

plot(Pstate)
}
}

