% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pstate.R
\name{Pstate}
\alias{Pstate}
\title{State probabilities filtering function.}
\usage{
Pstate(spec, theta, y)
}
\arguments{
\item{spec}{Model specification of class \code{\link{MSGARCH_SPEC}} created with \code{\link{create.spec}}.}

\item{theta}{Vector (of size d) or matrix (of size M x d) of parameter estimates.}

\item{y}{Vector (of size T) of observations.}
}
\value{
Filtered state probabilities (array of size T x M x K).
}
\description{
Method returning the filtered state probabilities.
}
\details{
If a matrix of parameter estimates is given, each parameter estimates is evaluated individually.
}
\examples{
data("sp500ret")

spec = MSGARCH::create.spec(model = c("sGARCH","sGARCH"), distribution = c("norm","norm"),
                             do.skew = c(FALSE,FALSE), do.mix = FALSE, do.shape.ind = FALSE) 

Pstate  = MSGARCH::Pstate(spec = spec, theta = spec$theta0, y = sp500ret)
}

