% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk.R
\name{Risk}
\alias{Risk}
\alias{Risk.MSGARCH_SPEC}
\alias{Risk.MSGARCH_ML_FIT}
\alias{Risk.MSGARCH_MCMC_FIT}
\title{Value-at-Risk and Expected-shortfall.}
\usage{
Risk(object, ...)

\method{Risk}{MSGARCH_SPEC}(object, par, data, alpha = c(0.01, 0.05),
  nahead = 1L, do.es = TRUE, do.its = FALSE, do.cumulative = FALSE,
  ctr = list(), ...)

\method{Risk}{MSGARCH_ML_FIT}(object, newdata = NULL, alpha = c(0.01,
  0.05), do.es = TRUE, do.its = FALSE, nahead = 1L,
  do.cumulative = FALSE, ctr = list(), ...)

\method{Risk}{MSGARCH_MCMC_FIT}(object, newdata = NULL, alpha = c(0.01,
  0.05), do.es = TRUE, do.its = FALSE, nahead = 1L,
  do.cumulative = FALSE, ctr = list(), ...)
}
\arguments{
\item{object}{Model specification of class \code{MSGARCH_SPEC}
created with \code{\link{CreateSpec}} or fit object of type \code{MSGARCH_ML_FIT}
created with \code{\link{FitML}} or \code{MSGARCH_MCMC_FIT}
created with \code{\link{FitMCMC}}.}

\item{...}{Not used. Other arguments to \code{Risk}.}

\item{par}{Vector (of size d) or matrix (of size \code{nmcmc} x d) of parameter estimates
where d must have
the same length as the default parameters of the specification.}

\item{data}{Vector (of size T) of observations.}

\item{alpha}{Vector (of size R) of Value-at-risk and Expected-shortfall levels.\cr
(Default: \code{alpha = c(0.01, 0.05)})}

\item{nahead}{Scalar indicating the number of step-ahead evaluation. (Default: \code{nahead = 1L}). Not used when
\code{do.its = TRUE} as it only returns in-sample one-step ahead risk measures.}

\item{do.es}{Logical indicating if Expected-shortfall is also calculated.
(Default: \code{do.es = TRUE})}

\item{do.its}{Logical indicating if the in-sample risk estimators are returned.
(Default: \code{do.its = FALSE}).}

\item{do.cumulative}{Logical indicating if the risk measures are computed on the 
cumulative simulations (typically log-returns, as they can be aggregated).
 Only available for \code{do.its = FALSE}. (Default: \code{do.cumulative = FALSE})}

\item{ctr}{A list of control parameters:
\itemize{
\item \code{nmesh} (integer >= 0) : Number of points for density
 evaluation. (Default: \code{nmesh = 1000L})
\item \code{nsim} (integer >= 0) :
Number indicating the number of simulation done for estimation of the
density at \code{nahead > 1}. (Default: \code{nsim = 10000L})
}}

\item{newdata}{Vector (of size T*) of new observations. (Default \code{newdata = NULL})}
}
\value{
A list of class \code{MSGARCH_RISK} with the following elements:
\itemize{
\item \code{VaR}:\cr
If \code{do.its = FALSE}: Value-at-Risk at \code{t = T + T* + 1, ... ,t = T + T* + nahead} at the
chosen levels (matrix of size \code{nahead} x R).\cr
If \code{do.its = TRUE}: In-sample Value-at-Risk at the chosen levels (Matrix of size (T + T*) x R).
\item \code{ES}:\cr
If \code{do.its = FALSE}: Expected-shortfall at \code{t = T + T* + 1, ... ,t = T + T* + nahead} at the
chosen levels (matrix of size \code{nahead} x R).\cr
If \code{do.its = TRUE}: In-sample Expected-shortfall at the chosen levels (Matrix of size (T + T*) x R).
}
The \code{MSGARCH_RISK} contains the \code{plot} method.
Note that the MCMC/Bayesian risk estimator can take long time to calculate
depending on the size of the MCMC chain.
}
\description{
Method returning the Value-at-Risk and Expected-shortfall risk measures.
}
\details{
If a matrix of MCMC posterior draws is given, the
Bayesian Value-at-Risk and Expected-shortfall are calculated.
Two or more step ahead risk measures are estimated via simulation of \code{nsim} paths up to
\code{t = T + T* + nahead}.
If \code{do.its = FALSE}, the risk estimators at \code{t = T + T* + 1, ... ,t = T + T* + nahead}
are computed. \code{do.cumulative = TRUE} indicate the function to compute the risk meausre 
over aggregated period up to \code{nahead} period using the \code{cumsum} function on the simulated data.
}
\examples{
# create specification
spec <- CreateSpec()

# load data
data("SMI", package = "MSGARCH")

# risk from specification
par <- c(0.1, 0.1, 0.8, 0.2, 0.1, 0.8, 0.99, 0.01)
set.seed(1234)
risk <- Risk(object = spec, par = par, data = SMI, nahead = 5L)
head(risk)
plot(risk)

# risk from ML fit
fit <- FitML(spec = spec, data = SMI)
set.seed(1234)
risk <- Risk(object = fit, nahead = 5L)
head(risk)
plot(risk)

\dontrun{
# risk from MCMC fit
set.seed(1234)
fit <- FitMCMC(spec = spec, data = SMI)
risk <- Risk(object = fit, nahead = 5L)
head(risk)
plot(risk)
}
}
