% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnd.R
\name{rnd}
\alias{rnd}
\title{Random draws at \code{t = T + 1} simulation method.}
\usage{
rnd(object, n, theta, y)
}
\arguments{
\item{object}{Model specification of class \code{MSGARCH_SPEC} created with \code{\link{create.spec}}
or fit object of type \code{MSGARCH_MLE_FIT} created with \code{\link{fit.mle}} or \code{MSGARCH_BAY_FIT}
created with \code{\link{fit.bayes}}.}

\item{n}{Number of random draws to be generated.}

\item{theta}{Vector (of size d) or matrix (of size M x d) of parameter estimates (not require when using a fit object).}

\item{y}{Vector (of size T) of observations (not require when using a fit object).}
}
\value{
A list of class \code{MSGARCH_RND} containing  two components:
\itemize{
\item \code{draws}: vector (of size n) or matrix (of size M x n) of simulated draws at \code{t = T + 1}.
\item \code{state}: vector (of size n) or matrix (of size M x n) of simulated states at \code{t = T + 1}.
}
The \code{MSGARCH_RND} class contains the \code{summary} and \code{plot} method.
}
\description{
Method returning random draws at \code{t = T + 1}.
}
\examples{
\dontrun{
# load data
data("sp500ret")

# create model specification
spec = MSGARCH::create.spec() 

# fit the model on the data with ML estimation using DEoptim intialization
set.seed(123)
fit = MSGARCH::fit.mle(spec = spec, y = sp500ret)
 
# generate random draws
set.seed(123)
rnd = MSGARCH::rnd(object = fit, n = 1000)

plot(rnd)

summary(rnd)
}
}

