% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{Simulation function.}
\usage{
sim(spec, n, theta, burnin = 500, do.state = FALSE)
}
\arguments{
\item{spec}{Model specification of class \code{MSGARCH_SPEC} created with \code{\link{create.spec}}.}

\item{n}{Simulation length.}

\item{theta}{Vector (of size d) or matrix (of size M x d) of parameter estimates.}

\item{burnin}{(integer >= 0) Burnin period discarded (first simulation draws). (default: \code{burnin = 500})}

\item{do.state}{Boolean  indicating if the simulated state are also output. (default: \code{log = TRUE})}
}
\value{
A list of class \code{MSGARCH_SIM} containing one or two components.
\itemize{
\item \code{draws}: vector (of size n) or matrix (of size M x n) of simulated draws.
\item \code{state}: vector (of size n) or matrix (of size M x n) of simulated states.
 The \code{state} value appear only if \code{do.state = TRUE}.
}
}
\description{
Method returning a simulated process.
}
\details{
If a matrix of parameter estimates is given, each parameter estimates is evaluated individually.
}
\examples{
spec = MSGARCH::create.spec(model = c("sGARCH","sGARCH"), distribution = c("norm","norm"),
                            do.skew = c(FALSE,FALSE), do.mix = FALSE, do.shape.ind = FALSE) 

y = MSGARCH::sim(spec = spec, n = 1000, theta = spec$theta0, burnin = 500, do.state = TRUE)

}

