\name{sp500ret}
\docType{data}
\alias{sp500ret}
\title{Standard and poors 500 closing Value log return}
\description{
The S&P500 index closing value log return from 1987-03-10 to 2009-01-30 from
Yahoo Finance \url{https://ca.finance.yahoo.com/}.
}
\usage{data("sp500ret")}
\format{Vector containing 5523 observations.}
\source{Yahoo Finance \url{https://ca.finance.yahoo.com/}}
\keyword{datasets}
