% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmat.R
\name{transmat}
\alias{transmat}
\title{Transition Matrix.}
\usage{
transmat(object, theta)
}
\arguments{
\item{object}{Model specification of class \code{MSGARCH_SPEC} created with \code{\link{create.spec}}
or fit object of type \code{MSGARCH_MLE_FIT} created with \code{\link{fit.mle}}.}

\item{theta}{Vector (of size d) of parameter estimates (not require when using a fit object).}
}
\value{
A matrix (of size K x K) in the case of a Markov-Switching model
 or a vector (of size K) in the case of a Mixture model.
}
\description{
Method returning the transition matrix.
}
\examples{
\dontrun{
# load data
data("sp500ret")

# create model specification
spec = MSGARCH::create.spec() 

# fit the model on the data with ML estimation using DEoptim intialization
set.seed(123)
fit = MSGARCH::fit.mle(spec = spec, y = sp500ret)

# run pdf method in-sample
transmat.mle = MSGARCH::transmat(fit)

print(transmat.mle)
}
}

